refactor: modularization of extract engine - further separation with extractors/*
